+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.0.1-dev-9-feature/DHP-13-16-save-survey
+- Saving and loading survey questions
+
+## 0.0.1-dev-9-feature/DHP-14-drag-drop-questions-2
+- Drag and drop added to the questions for sorting
+- Sorting fixed.
+
+## 0.0.1-dev-9-feature/DHP-20-question-pages
+- Question component added.
+- Page manager updated.
+- Questions per page added.
+
+## 0.0.1-dev-8-feature/DHP-12-options-url
+- Options URL added to field types.
+
 ## 0.0.1-dev-7-feature/DHP-8-field-types
 - Field type component added